split iota-identity - add iota-identity-core
