+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
   "identity-did",
   "identity-diff",
   "identity-iota",
+  "identity-iota-core",
 
   "examples",
 ]
