+++ identity-iota-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "identity-iota"
+name = "identity-iota-core"
 version = "0.5.0-dev.4"
 authors = ["IOTA Stiftung"]
 edition = "2021"
@@ -8,23 +8,12 @@ keywords = ["iota", "tangle", "identity"]
 license = "Apache-2.0"
 readme = "../README.md"
 repository = "https://github.com/iotaledger/identity.rs"
-description = "An IOTA Tangle intergration for the identity-rs library."
+description = ""
 
 [dependencies]
-async-trait = { version = "0.1", default-features = false }
-bee-rest-api = { version = "0.1.7", default-features = false }
-brotli = { version = "3.3", default-features = false, features = ["std"] }
-form_urlencoded = { version = "1.0" }
-futures = { version = "0.3" }
 identity-core = { version = "=0.5.0-dev.4", path = "../identity-core" }
-identity-credential = { version = "=0.5.0-dev.4", path = "../identity-credential" }
 identity-did = { version = "=0.5.0-dev.4", path = "../identity-did" }
-identity-iota-core = { version = "=0.5.0-dev.4", path = "../identity-iota-core" }
-itertools = { version = "0.10" }
 lazy_static = { version = "1.4", default-features = false }
-log = { version = "0.4", default-features = false }
-num-derive = { version = "0.3", default-features = false }
-num-traits = { version = "0.2", default-features = false, features = ["std"] }
 serde = { version = "1.0", default-features = false, features = ["std", "derive"] }
 strum = { version = "0.21", features = ["derive"] }
 thiserror = { version = "1.0", default-features = false }
@@ -35,19 +24,10 @@ rev = "7f5ae8f5b27f5948d8e1b4717419cc1321d862da"
 features = ["tls"]
 default-features = false
 
-[target.'cfg(all(target_arch = "wasm32", not(target_os = "wasi")))'.dependencies.iota-client]
-git = "https://github.com/iotaledger/iota.rs"
-rev = "7f5ae8f5b27f5948d8e1b4717419cc1321d862da"
-default-features = false
-features = ["wasm"]
-
 [dependencies.iota-crypto]
 version = "0.7"
 default-features = false
 features = ["blake2b"]
-
-[dev-dependencies]
-tokio = { version = "1.15", default-features = false, features = ["macros"] }
 
 [features]
 default = ["async"]
