+++ identity-iota-core/README.md
@@ -1,0 +1,1 @@
+The IOTA Identity Library
