+++ identity-iota-core/src/did/macros.rs
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 /// Creates a new IOTA DID from a `public` key and optional `network`.
