+++ identity-iota-core/src/tangle/traits.rs
@@ -1,0 +1,17 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::did::IotaDID;
+use crate::message::MessageId;
+
+pub trait TangleRef {
+  fn did(&self) -> &IotaDID;
+
+  fn message_id(&self) -> &MessageId;
+
+  fn set_message_id(&mut self, message_id: MessageId);
+
+  fn previous_message_id(&self) -> &MessageId;
+
+  fn set_previous_message_id(&mut self, message_id: MessageId);
+}
