+++ identity-iota-core/src/types/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub use network::Network;
+pub use network::NetworkName;
+
+mod network;
