+++ identity-iota-core/src/types/network.rs
