+++ identity-iota/src/chain/document_history.rs
@@ -1,26 +1,26 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 use std::collections::HashSet;
 use std::ops::Deref;
 
+use identity_iota_core::did::IotaDID;
+use identity_iota_core::diff::DiffMessage;
+use identity_iota_core::document::IotaDocument;
+use identity_iota_core::message::MessageId;
+use identity_iota_core::tangle::TangleRef;
 use serde;
 use serde::Deserialize;
 use serde::Serialize;
 
 use crate::chain::DiffChain;
 use crate::chain::IntegrationChain;
-use crate::did::IotaDID;
-use crate::diff::DiffMessage;
-use crate::document::IotaDocument;
 use crate::document::ResolvedIotaDocument;
 use crate::error::Result;
 use crate::tangle::Client;
 use crate::tangle::Message;
 use crate::tangle::MessageExt;
-use crate::tangle::MessageId;
 use crate::tangle::MessageIndex;
-use crate::tangle::TangleRef;
 
 /// A DID Document's history and current state.
 // ChainHistory<T> is not stored directly due to limitations on exporting generics in Wasm bindings.
