+++ identity-iota/src/document/resolved_iota_document.rs
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 use core::fmt::Debug;
@@ -6,18 +6,17 @@ use core::fmt::Display;
 use core::fmt::Formatter;
 use core::fmt::Result as FmtResult;
 
+use identity_core::convert::FmtJson;
+use identity_iota_core::did::IotaDID;
+use identity_iota_core::diff::DiffMessage;
+use identity_iota_core::document::IotaDocument;
+use identity_iota_core::message::MessageId;
+use identity_iota_core::message::MessageIdExt;
+use identity_iota_core::tangle::TangleRef;
 use serde::Deserialize;
 use serde::Serialize;
 
-use identity_core::convert::FmtJson;
-
-use crate::did::IotaDID;
-use crate::diff::DiffMessage;
-use crate::document::IotaDocument;
 use crate::error::Result;
-use crate::tangle::MessageId;
-use crate::tangle::MessageIdExt;
-use crate::tangle::TangleRef;
 
 /// An IOTA DID document resolved from the Tangle. Represents an integration chain message possibly
 /// merged with one or more diff messages.
@@ -58,7 +57,7 @@ impl ResolvedIotaDocument {
   /// Fails if the merge operation or signature verification on the diff fails.
   pub fn merge_diff_message(&mut self, diff_message: &DiffMessage) -> Result<()> {
     self.document.merge_diff(diff_message)?;
-    self.diff_message_id = diff_message.message_id;
+    self.diff_message_id = *diff_message.message_id();
 
     Ok(())
   }
