+++ identity-iota/src/tangle/client.rs
@@ -8,25 +8,25 @@ use iota_client::Client as IotaClient;
 use iota_client::Error as IotaClientError;
 
 use identity_core::convert::ToJson;
+use identity_iota_core::did::IotaDID;
+use identity_iota_core::diff::DiffMessage;
+use identity_iota_core::document::IotaDocument;
+use identity_iota_core::message::MessageId;
+use identity_iota_core::tangle::TangleRef;
+use identity_iota_core::types::Network;
 
 use crate::chain::ChainHistory;
 use crate::chain::DiffChain;
 use crate::chain::DocumentChain;
 use crate::chain::DocumentHistory;
 use crate::chain::IntegrationChain;
-use crate::did::IotaDID;
-use crate::diff::DiffMessage;
-use crate::document::IotaDocument;
 use crate::document::ResolvedIotaDocument;
 use crate::error::Error;
 use crate::error::Result;
 use crate::tangle::ClientBuilder;
 use crate::tangle::DIDMessageEncoding;
 use crate::tangle::Message;
-use crate::tangle::MessageId;
-use crate::tangle::Network;
 use crate::tangle::Receipt;
-use crate::tangle::TangleRef;
 use crate::tangle::TangleResolve;
 
 /// Client for performing IOTA Identity operations on the Tangle.
