+++ identity-iota/src/tangle/message/compression_brotli.rs
@@ -30,8 +30,7 @@ pub(crate) fn decompress_brotli<T: AsRef<[u8]> + ?Sized>(input: &T) -> Result<Ve
 mod test {
   use identity_core::convert::ToJson;
   use identity_core::crypto::KeyPair;
-
-  use crate::document::IotaDocument;
+  use identity_iota_core::document::IotaDocument;
 
   use super::*;
