+++ identity-iota/src/tangle/mod.rs
@@ -6,8 +6,6 @@
 pub use iota_client::bee_message::Error as BeeMessageError;
 #[doc(inline)]
 pub use iota_client::bee_message::Message;
-#[doc(inline)]
-pub use iota_client::bee_message::MessageId;
 
 pub use self::client::Client;
 pub use self::client_builder::ClientBuilder;
@@ -16,23 +14,18 @@ pub(crate) use self::message::pack_did_message;
 pub use self::message::DIDMessageEncoding;
 pub use self::message::DIDMessageVersion;
 pub use self::message::MessageExt;
-pub use self::message::MessageIdExt;
 pub use self::message::MessageIndex;
 pub use self::message::TryFromMessage;
-pub use self::network::Network;
-pub use self::network::NetworkName;
 pub use self::publish::PublishType;
 pub use self::receipt::Receipt;
 pub use self::resolver::Resolver;
 pub use self::resolver::ResolverBuilder;
-pub use self::traits::TangleRef;
 pub use self::traits::TangleResolve;
 
 mod client;
 mod client_builder;
 mod explorer;
 mod message;
-mod network;
 mod publish;
 mod receipt;
 mod resolver;
