+++ identity-iota/src/tangle/traits.rs
@@ -1,22 +1,10 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::did::IotaDID;
+use identity_iota_core::did::IotaDID;
+
 use crate::document::ResolvedIotaDocument;
 use crate::error::Result;
-use crate::tangle::MessageId;
-
-pub trait TangleRef {
-  fn did(&self) -> &IotaDID;
-
-  fn message_id(&self) -> &MessageId;
-
-  fn set_message_id(&mut self, message_id: MessageId);
-
-  fn previous_message_id(&self) -> &MessageId;
-
-  fn set_previous_message_id(&mut self, message_id: MessageId);
-}
 
 // TODO: remove TangleResolve with ClientMap refactor?
 #[async_trait::async_trait(?Send)]
